fix: delete procrustes.py
